added function call in db attributes of repos
organized backend main structure
